[newjersey/doula-pm#428] Add typechecking to cypress test files
Previously, a `tsconfig.json` was defined in `/cypress`, but not actually used. Type violations in e.g. `fillAndDownloadApplication.cy.ts` would not fail `npm run typecheck`.

This commit wires up type checking for cypress test files. As part of doing so, this commit refactors to break off `TestField` definitions and utility functions in `sharedTests.ts`, into a separate `src/app/form/_utils/testUtils/testFields.ts` file.

Some context for breaking up `sharedTests.ts` is that there are three potential sources of the `expect()` function:
1. cypress (chai)
2. vitest (or jest, if using instead)
3. testing-library/jest-dom, which provides additional functions like `toHaveAccessibleDescription()` on `expect()`

Two separate processes run for e2e testing. The application process that spins up the webapp, and the cypress process that runs the test against the webapp. For this cypress process, e.g. code executed in `fillAndDownloadApplication.cy.ts`, the `expect` is a cypress chai expect, with syntax like `expect(pageCount).to.equal(39);`.

Previously, `fillAndDownloadApplication.cy.ts` would `import { type TestField } from "@/app/form/_utils/testUtils/sharedTests";`. However, `sharedTests.ts` also contains code like `expect(input).toHaveFocus();`. These `expect()`s are meant to be run in the unit test process, which uses vitest + testing-library/jest-dom definitions of `expect()`. Since the cypress process uses the cypress global expect, type checking complains that if one were to run the code in that file, `.toHaveFocus()` another functions don't actually exist on the cypress `expect()`. Runtime-wise this is not a problem because we don't run that code during cypress testing, we just import the file in order to use the `TestField` type definition.

This commit thus separates out code relating to `TestField`, away from the testing utility function code. `fillAndDownloadApplication.cy.ts` now imports from `src/app/form/_utils/testUtils/testFields.ts` which does not contain any `expect()` statements.

This commit contributes to newjersey/doula-pm#428.
