+++ src/app/form/(formSteps)/business/3/testFields.ts
@@ -1,4 +1,4 @@
-import { createTestField, type TestField } from "@/app/form/_utils/testUtils/sharedTests";
+import { createTestField, type TestField } from "@/app/form/_utils/testUtils/testFields";
 
 const noHasBeenExcludedFromMedicaid: TestField = createTestField({
   name: "No",
