+++ src/app/form/(formSteps)/business/4/testFields.ts
@@ -2,7 +2,7 @@ import {
   createTestField,
   createTestFields,
   type TestField,
-} from "@/app/form/_utils/testUtils/sharedTests";
+} from "@/app/form/_utils/testUtils/testFields";
 
 const currentYear: number = new Date().getFullYear();
