+++ src/app/form/(formSteps)/insurance/2/testFields.ts
@@ -2,7 +2,7 @@ import {
   createTestField,
   createTestFields,
   type TestField,
-} from "@/app/form/_utils/testUtils/sharedTests";
+} from "@/app/form/_utils/testUtils/testFields";
 
 const insuranceAddressGroupName =
   "Insurance address This is the office location of your insurance carrier.";
