+++ src/app/form/(formSteps)/legal/1/testFields.ts
@@ -1,4 +1,4 @@
-import { createTestField, type TestField } from "@/app/form/_utils/testUtils/sharedTests";
+import { createTestField, type TestField } from "@/app/form/_utils/testUtils/testFields";
 
 export const noIsEmployedByNj: TestField = createTestField({
   name: "No",
