+++ src/app/form/(formSteps)/screening/3/testFields.ts
@@ -1,4 +1,4 @@
-import { createTestField, type TestField } from "@/app/form/_utils/testUtils/sharedTests";
+import { createTestField, type TestField } from "@/app/form/_utils/testUtils/testFields";
 
 const noHaveOtherBusinessOwnerNextYear = createTestField({
   name: "No",
